Refactor tests so that the status can use the auction_options fixture Change test to work with min/max auction
