+++ auction-deploy/tests/test_deploy.py
@@ -10,33 +10,6 @@ from auction_deploy.core import (
     missing_whitelisted_addresses,
 )
 
-RELEASE_TIMESTAMP_OFFSET = 3600 * 24 * 180
-
-
-@pytest.fixture
-def release_timestamp(web3):
-    """release timestamp used for DepositLocker contract"""
-    now = web3.eth.getBlock("latest").timestamp
-    return now + RELEASE_TIMESTAMP_OFFSET
-
-
-@pytest.fixture
-def auction_options(release_timestamp):
-    start_price = 1
-    auction_duration = 2
-    minimal_number_of_participants = 3
-    maximal_number_of_participants = 4
-
-    contract_options = AuctionOptions(
-        start_price=start_price,
-        auction_duration=auction_duration,
-        minimal_number_of_participants=minimal_number_of_participants,
-        maximal_number_of_participants=maximal_number_of_participants,
-        release_timestamp=release_timestamp,
-    )
-
-    return contract_options
-
 
 @pytest.fixture
 def deployed_contracts(web3, auction_options):
